requirements.txt
chatlas>=0.15.1
python_dotenv
pandas
pytest
requests
shiny
